feat: add init, update, quickstart subcommands (#9)
* add black, isort as requirements

* add prettify_python to tidy-up python files

* remove compile option from babelize command

* remove the checklist

* make isort quiet

* add OutputDirExistsError

* use pathlib

* change return value of babelize command to be new folder

* remove clobber option to babelize

* add rebabelize command

* better error handling with the babelize command

* reformat for black-styling

* remove lint

* remove py-scripting requirement

* remove unused data_files function

* fix output_dir bug in rebabelize

* init git repository and install versioneer

* add versioneer, gitpython as requirement

* remove extra cookiecutter files

* remove 2.7 builds, add 3.8 builds

* change plugin.yaml to babel.yaml

* remove rebabelize, bmi-render commands

* add MANIFEST file to template

* add METADATA attribute to Bmi classes

* add requirements file for testing

* clean up Travis config file

* remove model_metadata dependency

* include requirements.txt in MANIFEST

* run versioneer with subprocess.call

* link package data folder to meta folder

* remove distutils, just use setuptools

* test babelize subcommands

* add some documentation to the README, convert to rst

* remove markdown version of the README

* fix bulleted list in README

* remove unused bmi_register key

* add more documentation

* change PluginMetadata to BabelMetadata

* fix typo

* add format method to BabelMetadata

* add make and quickstart subcommands to babelize

* move main_babelize into cli.py

* test new subcommands

* only test for python 3.8

* reformat for black

* combine make subcommand into quickstart --batch

* combine make subcommand into quickstart --batch

* add docs that use babelize quickstart

+++ babelizer/data/hooks/post_gen_project.py
@@ -32,7 +32,7 @@ def clean_folder(folderpath, keep=None):
 
     folderpath = PROJECT_DIRECTORY / folderpath
     for fname in folderpath.glob("*"):
-        if str(fname.resolve()) not in keep:
+        if not fname.is_dir() and str(fname.resolve()) not in keep:
             fname.unlink()
 
     try:
@@ -108,10 +108,15 @@ if __name__ == "__main__":
     if "Not open source" == "{{ cookiecutter.open_source_license }}":
         remove_file("LICENSE")
 
+    datadir = Path("meta")
+    package_datadir = Path("pymt_{{cookiecutter.plugin_name}}") / "data"
+    if not package_datadir.exists():
+        package_datadir.symlink_to(".." / datadir, target_is_directory=True)
+
 {% for entry_point in cookiecutter.entry_points.split(',') %}
     {%- set pymt_class = entry_point.split('=')[0] %}
     write_api_yaml(
-        Path("meta", "{{ pymt_class }}"),
+        datadir / "{{ pymt_class }}",
         language="{{ cookiecutter.language }}",
         plugin_class="{{ pymt_class }}",
         plugin_name="{{ cookiecutter.plugin_name }}",
